+++ docs/pallas/CHANGELOG.md
@@ -41,6 +41,10 @@ Remember to align the itemized text with the first line of an item within a list
   * {class}`jax.experimental.pallas.tpu.CompilerParams` now defaults
     ``use_tc_tiling_on_sc`` to True for SparseCore kernels.
 
+* Deprecations
+
+  * Deprecated `pltpu.HOST` and `pltpu.MemorySpace.HOST` in favor of `pl.HOST`.
+
 ### Mosaic GPU
 
 * New features